Move Container classes into container_classes to avoid circular imports
