+++ lib/galaxy/tools/deps/dockerfiles.py
@@ -3,7 +3,7 @@ import os
 
 from ..deps import commands
 from ..deps import docker_util
-from ..deps.containers import docker_cache_path
+from ..deps.container_classes import docker_cache_path
 from ..deps.requirements import parse_requirements_from_xml
 from ...tools import loader_directory
